% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_dashboard.R
\name{deprecated-rjdqa}
\alias{deprecated-rjdqa}
\alias{sa_dashboard}
\title{Deprecated functions}
\usage{
sa_dashboard(x, n_recent_obs = 24)
}
\arguments{
\item{x}{a seasonal adjustment model made by 'RJDemetra' (object of class \code{"SA"}).}

\item{n_recent_obs}{number of observation in the recent history panel (see details). By default \code{n_recent_obs = 24} (last 2 years for monthly data).}
}
\description{
Use [sc_dashboard] instead of [sa_dashboard()].
}
