% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_export.R
\name{export_xlsx.mQA_matrix}
\alias{export_xlsx.mQA_matrix}
\title{Export des objets mQA_matrix dans des fichiers Excel}
\usage{
\method{export_xlsx}{mQA_matrix}(x, export_dir = "./",
  layout_file = c("ByComponent", "ByQRMatrix"), file_extension = c(".xls",
  ".xlsx"), layout = c("all", "modalities", "values", "combined"), ...)
}
\arguments{
\item{x}{objet de type \code{mQA_matrix} à exporter.}

\item{export_dir}{dossier d'export des résultats.}

\item{layout_file}{paramètre d'export. Par défaut (\code{layout_file = "ByComponent"}) on a un fichier Excel par composante
de la matrice bilan qualité (matrice des modalités ou des valeurs) dont chaque feuille correspond à un bilan qualité. Pour avoir
un fichier par bilan qualité dont chaque feuille correspond à la composante exportée, utiliser \code{layout_file = "ByQRMatrix"}.}

\item{file_extension}{extension des fichiers (\code{".xls"} ou \code{".xlsx"}).}

\item{layout}{composantes du bilan à exporter : voir \code{\link{export_xlsx.QA_matrix}} .}

\item{...}{autres paramètres de la fonction \code{\link{export_xlsx.QA_matrix}}.}
}
\description{
Permet d'exporter dans des fichiers Excel une liste de bilan qualité
}
\seealso{
Other QA_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QA_matrix}},
  \code{\link{extract_QA}}, \code{\link{extract_score}},
  \code{\link{rbind.QA_matrix}}, \code{\link{sort}},
  \code{\link{weighted_score}}
}
