% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_export.R
\name{export_xlsx.mQR_matrix}
\alias{export_xlsx.mQR_matrix}
\title{Export mQR_matrix objects to an Excel file}
\usage{
\method{export_xlsx}{mQR_matrix}(x, export_dir = "./",
  layout_file = c("ByComponent", "ByQRMatrix"), file_extension = c(".xls",
  ".xlsx"), layout = c("all", "modalities", "values", "combined"), ...)
}
\arguments{
\item{x}{a \code{mQR_matrix} object.}

\item{export_dir}{folder to export the result.}

\item{layout_file}{export parameter. By default (\code{layout_file = "ByComponent"}) an Excel by component
of the quality report matrix (modalities and/or values matrix) whose each sheet correspond to a QR_matrix. \code{layout_file = "ByQRMatrix"} to export 
one file per QR_matrix.}

\item{file_extension}{the files' extension (\code{".xls"} or \code{".xlsx"}).}

\item{layout}{QR_matrix' component to export. By default \code{layout = "all"}: the modalities matrix
(\code{"modalities"}) and the values matrix (\code{"values"}) are export. To add 
the additional variables of the values matrix to the modalities matrix, use \code{layout = "combined"}.}

\item{...}{other parameters of the \code{\link{export_xlsx.QR_matrix}} function.}
}
\description{
Function to export a list of quality report matrix  (\code{mQR_matrix}) to an Excel file.
}
\seealso{
Other QR_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QR_matrix}},
  \code{\link{extract_QR}}, \code{\link{extract_score}},
  \code{\link{rbind.QR_matrix}}, \code{\link{sort}},
  \code{\link{weighted_score}}
}
