% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_export.R
\name{export_xlsx.QR_matrix}
\alias{export_xlsx.QR_matrix}
\title{Export QR_matrix objects to an Excel file}
\usage{
\method{export_xlsx}{QR_matrix}(x, layout = c("all", "modalities",
  "values", "combined"), create = TRUE, clear_sheet = TRUE,
  auto_format = TRUE, file_name, sheet_names, ...)
}
\arguments{
\item{x}{a \code{QR_matrix} object.}

\item{layout}{QR_matrix' component to export. By default \code{layout = "all"}: the modalities matrix
(\code{"modalities"}) and the values matrix (\code{"values"}) are export. To add 
the additional variables of the values matrix to the modalities matrix, use \code{layout = "combined"}.}

\item{create}{boolean indicating whether to create a new Excel file or not (\code{TRUE} by default).}

\item{clear_sheet}{boolean indicating whether to clear the Excel's sheets before exporting  (\code{TRUE} by default); 
usefull when the Excel file already exist.}

\item{auto_format}{boolean indicating whether to format the ouptut or not (\code{TRUE} by default).}

\item{file_name}{argument optionnel indiquant chemin vers le fichier à exporter. Si no spécifié
alors un fichier *export.xls* est créé dans le working directory.}

\item{sheet_names}{nom des feuilles Excel en sortie. Si non spécifié alors le nom correspond à la composante exportée.
Si le paramètre est spécifié alors les éventuelles feuilles contenant ces noms sont supprimées.}

\item{...}{autres paramètres non utilisés.}
}
\description{
Function to export a quality report matrix  (\code{QR_matrix}) to an Excel file.
}
\seealso{
Other QR_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.mQR_matrix}},
  \code{\link{extract_QR}}, \code{\link{extract_score}},
  \code{\link{rbind.QR_matrix}}, \code{\link{sort}},
  \code{\link{weighted_score}}
}
\concept{QR_matrix functions}
