% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_extract_functions.R
\encoding{UTF-8}
\name{extract_QA}
\alias{extract_QA}
\alias{extract_QA}
\alias{extract_QA.SA}
\alias{extract_QA}
\alias{extract_QA.default}
\title{Extract a quality assessment matrix}
\usage{
extract_QA(x, ...)

\method{extract_QA}{SA}(x, series_name, ...)

\method{extract_QA}{default}(x, sep = ";", dec = ",", ...)
}
\arguments{
\item{x}{a CSV file containing the diagnostics matrix or a \code{SA} object. By default a dialog box opens to select a CSV file.}

\item{...}{unused argument.}

\item{series_name}{The name of series for the \code{extract_QA.SA} method.
By default the name of the \code{x} parameter is used.}

\item{sep}{the field separator character in the CSV file (\code{sep = ";"} by default).}

\item{dec}{the character used in the file for decimal points (\code{dec = ","} by default).}
}
\value{
A \code{\link{QA_matrix}} object.
}
\description{
Functions to extracts a quality report assessment from a CSV file exported by JDemetra+ (or the JDemetra+ cruncher) or from a seasonal adjustment object create by the RJDemetra package.
}
\details{
The function allows to extract a quality assessment matrix from a CSV file containing all the diagnostics 
(usually the \emph{demetra_m.csv} file) or from a \code{SA} object create by the package \code{RJDemetra}).

The result is a \code{\link{QA_matrix}} object which is a list of three parameters:
* le paramètre \code{modalities} est un \code{data.frame} contenant un ensemble de variables sous forme catégorielle
  (Good, Uncertain, Bad, Severe).
* le paramètre \code{values} est un \code{data.frame} contenant les valeurs associées aux indicateurs présents dans
  \code{modalities} (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas
  de modalité (fréquence de la série et modèle ARIMA).
* le paramètre \code{score_formula} est initié à \code{NULL} : il contiendra la formule utilisée pour
  calculer le score (si le calcul est fait).
}
\examples{
\dontrun{
QA <- extract_QA()
QA
# To extract the modalities' matrix:
QA$modalities
# Ou :
QA[["modalities"]]
}
}
\seealso{
Other QA_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QA_matrix}},
  \code{\link{export_xlsx.mQA_matrix}},
  \code{\link{extract_score}},
  \code{\link{rbind.QA_matrix}}, \code{\link{sort}},
  \code{\link{weighted_score}}
}
