% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_extract_functions.R
\encoding{UTF-8}
\name{extract_QR}
\alias{extract_QR}
\alias{extract_QR.SA}
\alias{extract_QR.default}
\title{Extract a quality report matrix}
\usage{
extract_QR(x, ...)

\method{extract_QR}{SA}(x, series_name, ...)

\method{extract_QR}{default}(x, sep = ";", dec = ",", ...)
}
\arguments{
\item{x}{a CSV file containing the diagnostics matrix or a \code{SA} object. By default a dialog box opens to select a CSV file.}

\item{...}{unused argument.}

\item{series_name}{The name of series for the \code{extract_QR.SA} method.
By default the name of the \code{x} parameter is used.}

\item{sep}{the field separator character in the CSV file (\code{sep = ";"} by default).}

\item{dec}{the character used in the file for decimal points (\code{dec = ","} by default).}
}
\value{
A \code{\link{QR_matrix}} object.
}
\description{
Functions to extracts a quality report matrix from a CSV file exported by JDemetra+ (or the JDemetra+ cruncher) or from a seasonal adjustment object create by the RJDemetra package.
}
\details{
La fonction permet d'extraire un bilan qualité à partir d'un fichier csv contenant l'ensemble des
diagnostics (généralement fichier \emph{demetra_m.csv}).

Ce fichiers peut être obtenu en lançant le cruncher (\code{{cruncher}} ou \code{{cruncher_and_param}}) avec
l'ensemble des paramètres de base pour les paramètres à exporter et l'option \code{csv_layout = "vtable"} (par défaut)
pour le format de sortie des fichiers csv (option de \code{{cruncher_and_param}} ou de \code{{create_param_file}}
lors de la création du
fichier de paramètres).

Le résultat de cette fonction est un objet \code{\link{QR_matrix}} qui est une liste de trois paramètres :
* le paramètre \code{modalities} est un \code{data.frame} contenant un ensemble de variables sous forme catégorielle
  (Good, Uncertain, Bad, Severe).
* le paramètre \code{values} est un \code{data.frame} contenant les valeurs associées aux indicateurs présents dans
  \code{modalities} (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas
  de modalité (fréquence de la série et modèle ARIMA).
* le paramètre \code{score_formula} est initié à \code{NULL} : il contiendra la formule utilisée pour
  calculer le score (si le calcul est fait).
}
\examples{
\dontrun{
QR <- extract_QR()
QR
# To extract the modalities' matrix:
QR$modalities
# Ou :
QR[["modalities"]]
}
}
\seealso{
Other QR_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QR_matrix}},
  \code{\link{export_xlsx.mQR_matrix}},
  \code{\link{extract_score}},
  \code{\link{rbind.QR_matrix}}, \code{\link{sort}},
  \code{\link{weighted_score}}
}
\concept{QR_matrix functions}
