% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_dashboard.R
\name{plot.sa_dashboard}
\alias{plot.sa_dashboard}
\title{Plot a seasonal adjustment dashboard}
\usage{
\method{plot}{sa_dashboard}(x, main = "Seasonal Adjustment Dashboard",
  subtitle = "", raw_color = "#33A02C", sa_color = "#E31A1C",
  trend_color = "black", ...)
}
\arguments{
\item{x}{a \code{"sa_dashboard"} object}
}
\description{
Plot a seasonal adjustment dashboard
}
