% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_matrix_classes.R
\encoding{UTF-8}
\name{print.QA_matrix}
\alias{print.QA_matrix}
\alias{print.mQA_matrix}
\title{Printing QA_matrix and mQA_matrix}
\usage{
\method{print}{QA_matrix}(x, print_variables = TRUE,
  print_score_formula = TRUE, ...)

\method{print}{mQA_matrix}(x, score_statistics = TRUE, ...)
}
\arguments{
\item{x}{object of class code{mQA_matrix} or \code{mQA_matrix}.}

\item{print_variables}{logical, indicating whether or not the variables names should be printed.}

\item{print_score_formula}{logical, indicating whether or not the score formula should be printed.}

\item{...}{autres arguments non utilisés.}

\item{score_statistics}{logical, indicating whether or not the scores statistics should be printed.}
}
\description{
Print a QA_matrix or a mQA_matrix.
}
