% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_matrix_classes.R
\encoding{UTF-8}
\name{QA_matrix}
\alias{QA_matrix}
\alias{mQA_matrix}
\alias{is.QA_matrix}
\alias{is.mQA_matrix}
\title{Objets bilan qualité}
\usage{
QA_matrix(modalities = NULL, values = NULL, score_formula = NULL)

mQA_matrix(x = list(), ...)

is.QA_matrix(x)

is.mQA_matrix(x)
}
\arguments{
\item{modalities}{un \code{data.frame} contenant les modalités (Good, Bad, etc.)
associées aux variables.}

\item{values}{un \code{data.frame} contenant les valeurs (p-valeurs des tests, statistiques, etc.)
associées aux variables. Peut donc contenir plus de variables que le paramètre \code{modalities}.}

\item{score_formula}{formule utilisée pour calculer le score global (s'il existe).}

\item{x}{un objet de type \code{QA_matrix}, \code{mQA_matrix} ou une liste d'objets \code{QA_matrix}.}

\item{...}{des objets du même type que \code{x}.}
}
\description{
\code{QA_matrix} permet de créer un objet de type \code{QA_matrix} contenant un bilan qualité.
}
\details{
\code{mQA_matrix} permet de créer un objet de type \code{mQA_matrix} qui est une liste de bilans qualité (donc d'objets \code{QA_matrix}).

\code{is.QA_matrix} et \code{is.mQA_matrix} permettent de tester si un objet est un bilan qualité ou une liste de bilans qualité.

Un objet  de type \code{\link{QA_matrix}} est une liste de trois paramètres :
* le paramètre \code{modalities} est un \code{data.frame} contenant un ensemble de variables sous forme catégorielle
  (par défaut : Good, Uncertain, Bad, Severe).
* le paramètre \code{values}  est un \code{data.frame} contenant les valeurs associées aux indicateurs présents dans
  \code{modalities} (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas
  de modalité (e.g. : fréquence de la série, modèle ARIMA, etc.).
* le paramètre \code{score_formula} contient la formule utilisée pour calculer le score (une fois le calcul réalisé).
}
