% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_functions.R
\name{QR_var_manipulation}
\alias{QR_var_manipulation}
\alias{remove_indicators}
\alias{retain_indicators}
\title{Manipulate indicators of QR_matrix or mQR_matrix}
\usage{
remove_indicators(x, ...)

retain_indicators(x, ...)
}
\arguments{
\item{x}{\code{QR_matrix} or \code{mQR_matrix} objects.}

\item{...}{names of the indicators to remove (or retain).}
}
\description{
Functions to remove indicators (\code{remove_indicators}) or to retain indicators (\code{retain_indicators}) 
\code{QR_matrix} or \code{mQR_matrix}. The series' name ("series" columns) cannot be removed.
}
\examples{
\dontrun{
QR <- compute_score(extract_QR())
retain_indicators(QR,"score","m7") # To only retain the score and m7 statistics
retain_indicators(QR,c("score","m7")) #Equivalent
score(remove_indicator(QR,"score")) # To remove the score
}
}
\seealso{
Other var QR_matrix manipulation: \code{\link{add_indicator}},
  \code{\link{recode_indicator_num}}
}
\concept{var QR_matrix manipulation}
