% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_functions.R
\name{rbind.QA_matrix}
\alias{rbind.QA_matrix}
\title{Combiner par ligne des objets QA_matrix}
\usage{
\method{rbind}{QA_matrix}(..., check_formula = TRUE)
}
\arguments{
\item{...}{objets \code{QA_matrix} à combiner.}

\item{check_formula}{booléen indiquant s'il faut vérifier la cohérence dans les formule du calcul du score.
Par défaut \code{check_formula = TRUE} : la fonction renvoie une erreur si des scores sont calculés avec des formules différentes.
Si \code{check_formula = FALSE} alors il n'y a pas de vérification et le paramètre \code{score_formula} de l'objet
en sortie est \code{NULL}.}
}
\description{
Permet de combiner plusieurs objets \code{QA_matrix} en combinant par ligne les paramètres \code{modalities}
et \code{values}.
}
\examples{
\dontrun{
QA <- extract_QA()
QA1 <- compute_score(QA1, score_pond = c(m7 = 2, q = 3, qs_residual_sa_on_sa = 5))
QA2 <- compute_score(QA1, score_pond = c(m7 = 2, qs_residual_sa_on_sa = 5))
rbind(QA1, QA2) # Une erreur est renvoyée
rbind(QA1, QA2, check_formula = FALSE)
}
}
\seealso{
Other QA_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QA_matrix}},
  \code{\link{export_xlsx.mQA_matrix}},
  \code{\link{extract_QA}}, \code{\link{extract_score}},
  \code{\link{sort}}, \code{\link{weighted_score}}
}
\concept{QA_matrix functions}
