% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_functions.R
\name{sort}
\alias{sort}
\alias{sort.QA_matrix}
\alias{sort.mQA_matrix}
\title{Tri des objets QA_matrix et mQA_matrix}
\usage{
\method{sort}{QA_matrix}(x, decreasing = FALSE,
  sort_variables = "score", ...)

\method{sort}{mQA_matrix}(x, decreasing = FALSE,
  sort_variables = "score", ...)
}
\arguments{
\item{x}{objet de type \code{QA_matrix} ou \code{mQA_matrix}.}

\item{decreasing}{booléen indiquant si les bilans qualité doivent être triés par ordre croissant ou décroissant.
Par défaut le tri est effectué par ordre croissant.}

\item{sort_variables}{variables à utiliser pour le tri qui sont présentes dans les tables des modalités.}

\item{...}{autres paramètres de la fonction \code{\link[base]{order}} (non utilisés pour l'instant).}
}
\value{
L'objet en entrée avec les tables de bilan qualité triées.
}
\description{
Permet de trier les bilans qualité en fonction d'une ou plusieurs variables.
}
\examples{
\dontrun{
QA <- compute_score(extract_QA())
sort(QA, sort_variables = "score") #Pour trier par ordre croissant sur le score
}
}
\seealso{
Other QA_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QA_matrix}},
  \code{\link{export_xlsx.mQA_matrix}},
  \code{\link{extract_QA}}, \code{\link{extract_score}},
  \code{\link{rbind.QA_matrix}},
  \code{\link{weighted_score}}
}
\concept{QA_matrix functions}
