% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_functions.R
\name{weighted_score}
\alias{weighted_score}
\title{Calcul d'un score pondéré pour chaque observation}
\usage{
weighted_score(x, pond = 1)
}
\arguments{
\item{x}{objet de type \code{QA_matrix} ou \code{mQA_matrix}.}

\item{pond}{pondération à appliquer au score. Il peut s'agir d'un nombre, d'un vecteur de nombres, du nom
d'une des variables du bilan qualité ou d'une liste de pondérations pour les objets \code{mQA_matrix}.}
}
\value{
L'objet en entrée avec le score recalculé
}
\description{
Permet de pondérer un score déjà calculé en fonction des variables.
}
\examples{
\dontrun{
QR <- extract_QR()
QR <- compute_score(QR)
weighted_score(QR, 2) # Tous les scores sont multipliés par 2
}
}
\seealso{
Other QA_matrix functions: \code{\link{compute_score}},
  \code{\link{export_xlsx.QA_matrix}},
  \code{\link{export_xlsx.mQA_matrix}},
  \code{\link{extract_QA}}, \code{\link{extract_score}},
  \code{\link{rbind.QA_matrix}}, \code{\link{sort}}
}
