% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR_functions.R
\name{QA_var_manipulation}
\alias{QA_var_manipulation}
\alias{remove_indicators}
\alias{retain_indicators}
\title{Manipulate indicators of QA_matrix or mQA_matrix}
\usage{
remove_indicators(x, ...)

retain_indicators(x, ...)
}
\arguments{
\item{x}{\code{QA_matrix} or \code{mQA_matrix} objects.}

\item{...}{names of the indicators to remove (or retain).}
}
\description{
Functions to remove indicators (\code{remove_indicators}) or to retain indicators (\code{retain_indicators}) 
\code{QA_matrix} or \code{mQA_matrix}. The series' name ("series" columns) cannot be removed.
}
\examples{
\dontrun{
QA <- compute_score(extract_QA())
retain_indicators(QA,"score","m7") # To only retain the score and m7 statistics
retain_indicators(QA,c("score","m7")) #Equivalent
score(remove_indicator(QA,"score")) # To remove the score
}
}
\seealso{
Other var QA_matrix manipulation: \code{\link{add_indicator}},
  \code{\link{recode_indicator_num}}
}
